+++ ReversiTests/GameTests.swift
@@ -207,1031 +207,5 @@ class GameTests: XCTestCase {
             return state.thinking && state.turn == .light
         })
         wait(for: [lightThinkingExpectation], timeout: 3.0)
-    }
-    
-    func testInitialPhase() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(InitialPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-        
-        // .start()をディスパッチしたらStateが現在の状態に正しく初期化される。
-        let expectStateToBeInitialized = expectation(description: "State is initialized")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectStateToBeInitialized, inputChecker: { state in
-            // カウント数
-            guard state.diskCount[.dark] == 9 else { return false }
-            guard state.diskCount[.light] == 12 else { return false }
-            
-            // 思考中ではない
-            guard !state.thinking else { return false}
-            
-            // プレイヤーモード
-            guard state.playerModes[.dark] == .computer else { return false }
-            guard state.playerModes[.light] == .manual else { return false }
-            
-            // 白のターン
-            guard state.turn == .light else { return false }
-
-            // ボードの更新依頼が出る
-            guard state.boardUpdateRequest != nil else { return false }
-            
-            return true
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        
-        // また、次のフェーズは WaitForPlayer になる
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'WaitForPlayer'")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .waitForPlayer
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-        
-        game.dispatch(.start())
-        wait(for: [expectStateToBeInitialized, expectPhaseToBeWaitForPlayer], timeout: 3.0)
-    }
-    
-    func testWaitForPlayerPhase1() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(WaitForPlayerPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 黒のターンだが、黒はcomputerなのでThinkingPhaseへ遷移する
-        let expectPhaseToBeThinking = expectation(description: "Phase transits to 'Thinking'")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseToBeThinking, inputChecker: { state in
-            return state.phase.kind == .thinking
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-        
-        if let onEnter = WaitForPlayerPhase.onEnter(previousPhase: AnyPhase(InitialPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPhaseToBeThinking], timeout: 3.0)
-    }
-    
-    func testWaitForPlayerPhase2() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(WaitForPlayerPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 白のターンだが、白はmanualなのでそのまま待っていてもWaitForPlayerにとどまる
-        let expectPhaseNotToTransit = expectation(description: "Phase transits to nothing but 'WaitForPlayer'")
-        expectPhaseNotToTransit.isInverted = true
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseNotToTransit, inputChecker: { state in
-            return state.phase.kind != .waitForPlayer
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-        
-        if let onEnter = WaitForPlayerPhase.onEnter(previousPhase: AnyPhase(InitialPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPhaseNotToTransit], timeout: 1.0)
-        
-        // ボードのセルをタップすると、フェーズがPlaceDiskに移る
-        let expectPhaseToBePlaceDisk = expectation(description: "Phase transits to 'PlaceDisk'")
-        stateSubscriber.reset(expectPhaseToBePlaceDisk, inputChecker: { state in
-            guard let phase = PlaceDiskPhase.thisPhase(of: state) else { return false }
-            guard phase.x == 1 else { return false }
-            guard phase.y == 3 else { return false }
-            
-            return true
-        })
-        game.dispatch(.boardCellSelected(x: 1, y: 3))
-        wait(for: [expectPhaseToBePlaceDisk], timeout: 3.0)
-    }
-    
-    func testWaitForPlayerPhase3() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(WaitForPlayerPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 白のターンだが、白はmanual
-        if let onEnter = WaitForPlayerPhase.onEnter(previousPhase: AnyPhase(InitialPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-
-        // 白のプレイヤーモードをcomputerに変更すると
-        // ThinkingPhaseに移る
-        let expectPhaseToBeThinking = expectation(description: "Phase transits to 'Thinking'")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseToBeThinking, inputChecker: { state in
-            return state.phase.kind == .thinking
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-        
-        game.dispatch(.playerModeChanged(player: .light, mode: .computer))
-        wait(for: [expectPhaseToBeThinking], timeout: 3.0)
-    }
-    
-    func testWaitForPlayerPhase4() throws {
-        let board = try Board(restoringFromText: """
-            --------
-            --------
-            ----x---
-            ---xxx--
-            --xxxxx-
-            ---xxx--
-            ----x---
-            --------
-            """)
-        var state = State(board: board, turn: nil, playerModes: [.computer, .manual])
-        state.diskCount[.dark] = 13
-        state.diskCount[.light] = 0
-        state.phase = AnyPhase(WaitForPlayerPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // ターンがnilなのでGameOverへ遷移する
-        let expectPhaseToBeGameOver = expectation(description: "Phase transits to 'GameOver'")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseToBeGameOver, inputChecker: { state in
-            return state.phase.kind == .gameOver
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-        
-        if let onEnter = WaitForPlayerPhase.onEnter(previousPhase: AnyPhase(InitialPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPhaseToBeGameOver], timeout: 3.0)
-    }
-
-    func testThinkingPhase1() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(ThinkingPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 思考中になったらぐるぐるする
-        let expectThinking = expectation(description: "'thining' in state is true")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectThinking, inputChecker: { state in
-            return state.thinking
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-        
-        if let onEnter = ThinkingPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectThinking], timeout: 3.0)
-    }
-    
-    func testThinkingPhase2() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(ThinkingPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 思考中になり、しばらく待っていたら、ディスクを打つ（PlaceDiskフェーズへ遷移する）
-        // そしたらまた、思考中でなくなる。
-        var hasEverBeenThinking = false
-        let expectThinking = expectation(description: "'thining' in state becomes true and then false")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectThinking, inputChecker: { state in
-            if !hasEverBeenThinking && state.thinking {
-                hasEverBeenThinking = true
-            }
-            return hasEverBeenThinking && !state.thinking
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        
-        let expectPhaseToBePlaceDisk = expectation(description: "Computer automatically place the disk")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectPhaseToBePlaceDisk, inputChecker: { state in
-            return state.phase.kind == .placeDisk
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        if let onEnter = ThinkingPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectThinking, expectPhaseToBePlaceDisk], timeout: 3.0)
-    }
-    
-    func testThinkingPhase3() throws {
-        ThinkingPhase.thinkingDuration = 3.0
-        
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(ThinkingPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 思考中になってぐるぐるし始めたあと、
-        let expectThinking = expectation(description: "'thining' in state is true")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectThinking, inputChecker: { state in
-            return state.thinking
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        if let onEnter = ThinkingPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectThinking], timeout: 3.0)
-
-        // 思考中の間に黒のプレイヤーモードをmanualに変えたら
-        // WaitForPlayerフェーズに戻る
-        // そのとき思考中ではなくなる
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'WaitForPlayer'")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .waitForPlayer
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        let expectNotThinking = expectation(description: "'thining' in state is false")
-        stateSubscriber1.reset(expectNotThinking, inputChecker: { state in
-            return !state.thinking
-        })
-
-        game.dispatch(.playerModeChanged(player: .dark, mode: .manual))
-        wait(for: [expectPhaseToBeWaitForPlayer, expectNotThinking], timeout: 3.0)
-    }
-    
-    func testPlaceDiskPhase1() throws {
-        var board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(PlaceDiskPhase(x: 7, y: 5))
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 打てたらアニメーションさせるセルの情報を引き連れてPlacingDiskへ遷移
-        let cellChanges = try board.placeDisk(.light, atX: 7, y: 5)[...]
-        let expectPhaseToBePlacingDisk = expectation(description: "Phase transits to 'PlacingDisk'")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseToBePlacingDisk, inputChecker: { state in
-            guard let phase = PlacingDiskPhase.thisPhase(of: state) else { return false }
-            guard phase.cellChanges == cellChanges else { return false }
-            
-            return true
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-
-        if let onEnter = PlaceDiskPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPhaseToBePlacingDisk], timeout: 3.0)
-    }
-    
-    func testPlaceDiskPhase2() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(PlaceDiskPhase(x: 2, y: 2))
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 打てないところに打とうとしているので、
-        // WaitForPlayerフェーズに戻る
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'WaitForPlayer'")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .waitForPlayer
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-
-        if let onEnter = PlaceDiskPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPhaseToBeWaitForPlayer], timeout: 3.0)
-    }
-    
-    func testPlacingDiskPhase1() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---ooooo
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        let cellChanges = [
-            Board.CellChange(x: 7, y: 5, disk: .light),
-            Board.CellChange(x: 6, y: 5, disk: .light),
-            Board.CellChange(x: 5, y: 5, disk: .light),
-            Board.CellChange(x: 4, y: 5, disk: .light),
-        ]
-        state.phase = AnyPhase(PlacingDiskPhase(cellChanges: cellChanges[...]))
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 1つ目のセルの表示を更新する依頼が発行され、
-        let expectBoardUpdateRequest = expectation(description: "Request for board update")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectBoardUpdateRequest, inputChecker: { state in
-            if let request = state.boardUpdateRequest {
-                if case let .withAnimation(cellChange) = request.detail {
-                    if cellChange == cellChanges[0] {
-                        return true
-                    }
-                }
-            }
-            return false
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-
-        if let onEnter = PlacingDiskPhase.onEnter(previousPhase: AnyPhase(PlaceDiskPhase(x: 7, y: 5))) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectBoardUpdateRequest], timeout: 3.0)
-
-        // その依頼が完了するまでは、勝手に次の表示更新には進まずに
-        // そのままの状態でとどまる
-        let expectBoardUpdateRequestNotChange = expectation(description: "Request for board update")
-        expectBoardUpdateRequestNotChange.isInverted = true
-        stateSubscriber.reset(expectBoardUpdateRequestNotChange, inputChecker: { state in
-            if let request = state.boardUpdateRequest {
-                if case let .withAnimation(cellChange) = request.detail {
-                    if cellChange == cellChanges[0] {
-                        return false
-                    }
-                }
-            }
-            return true
-        })
-        wait(for: [expectBoardUpdateRequestNotChange], timeout: 1.0)
-    }
-
-    func testPlacingDiskPhase2() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---ooooo
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        let cellChanges = [
-            Board.CellChange(x: 7, y: 5, disk: .light),
-            Board.CellChange(x: 6, y: 5, disk: .light),
-            Board.CellChange(x: 5, y: 5, disk: .light),
-            Board.CellChange(x: 4, y: 5, disk: .light),
-        ]
-        state.phase = AnyPhase(PlacingDiskPhase(cellChanges: cellChanges[...]))
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 依頼を完了していけば、次々と依頼が発行されて、
-        // 最終的にNextTurnフェーズに遷移する
-        var requestedChanges = [Board.CellChange]()
-        gameState
-            .compactMap { (state) -> (requestId: UniqueIdentifier, cellChange: Board.CellChange)? in
-                if let request = state.boardUpdateRequest {
-                     if case let .withAnimation(cellChange) = request.detail {
-                        return (requestId: request.requestId, cellChange: cellChange)
-                    }
-                }
-                return nil
-            }
-            .removeDuplicates { $0.requestId == $1.requestId }
-            .sink { value in
-                requestedChanges.append(value.cellChange)
-                DispatchQueue.main.async {
-                    game.dispatch(.boardUpdated(requestId: value.requestId))
-                }
-            }
-            .store(in: &cancellables)
-        
-        let expectPhaseToBeNextTurn = expectation(description: "Phase transits to 'NextTurn'")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPhaseToBeNextTurn, inputChecker: { state in
-            return state.phase.kind == .nextTurn
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-
-        if let onEnter = PlacingDiskPhase.onEnter(previousPhase: AnyPhase(PlaceDiskPhase(x: 7, y: 5))) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPhaseToBeNextTurn], timeout: 3.0)
-        XCTAssertEqual(requestedChanges, cellChanges)
-    }
-    
-    func testNextTurnPhase1() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---ooooo
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.manual, .manual])
-        state.diskCount[.dark] = 9
-        state.diskCount[.light] = 12
-        state.phase = AnyPhase(NextTurnPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // ターンが黒に変わって、ディスク枚数が更新され、
-        // WaitForPlayerフェーズに遷移する
-        let expectTurnChange = expectation(description: "Changed to dark's turn")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectTurnChange, inputChecker: { state in
-            return state.turn == .dark
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-
-        let expectDiskCount = expectation(description: "Disk-counts change")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectDiskCount, inputChecker: { state in
-            return state.diskCount[.dark] == 6 && state.diskCount[.light] == 16
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'WaitForPlayer'")
-        let stateSubscriber3 = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .waitForPlayer
-        })
-        gameState.subscribe(stateSubscriber3)
-        stateSubscriber3.store(in: &cancellables)
-        
-        if let onEnter = NextTurnPhase.onEnter(previousPhase: AnyPhase(PlacingDiskPhase(cellChanges: [][...]))) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectTurnChange, expectDiskCount, expectPhaseToBeWaitForPlayer], timeout: 3.0)
-    }
-    
-    func testNextTurnPhase2() throws {
-        let board = try Board(restoringFromText: """
-            --------
-            --------
-            xooooooo
-            xxoooooo
-            xoxxoooo
-            -ooxoooo
-            --oooo--
-            --oooo--
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.manual, .manual])
-        state.diskCount[.dark] = 4
-        state.diskCount[.light] = 34
-        state.phase = AnyPhase(NextTurnPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // ターンが白に変わって、ディスク枚数が更新されるが、
-        // 白は打つところがないので、Passフェーズに遷移する
-        let expectTurnChange = expectation(description: "Changed to light's turn")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectTurnChange, inputChecker: { state in
-            return state.turn == .light
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-
-        let expectDiskCount = expectation(description: "Disk-counts change")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectDiskCount, inputChecker: { state in
-            return state.diskCount[.dark] == 7 && state.diskCount[.light] == 32
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'Pass'")
-        let stateSubscriber3 = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .pass
-        })
-        gameState.subscribe(stateSubscriber3)
-        stateSubscriber3.store(in: &cancellables)
-        
-        if let onEnter = NextTurnPhase.onEnter(previousPhase: AnyPhase(PlacingDiskPhase(cellChanges: [][...]))) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectTurnChange, expectDiskCount, expectPhaseToBeWaitForPlayer], timeout: 3.0)
-    }
-    
-    func testNextTurnPhase3() throws {
-        let board = try Board(restoringFromText: """
-            --------
-            --------
-            ----x---
-            ---xxx--
-            --xxxxx-
-            ---xxx--
-            ----x---
-            --------
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.manual, .manual])
-        state.diskCount[.dark] = 7
-        state.diskCount[.light] = 5
-        state.phase = AnyPhase(NextTurnPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // ターンが白に変わって、ディスク枚数が更新されるが、
-        // 白は打つところがなく、黒も打てないので、GameOverフェーズに遷移する
-        let expectTurnChange = expectation(description: "Changed to light's turn")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectTurnChange, inputChecker: { state in
-            return state.turn == .light
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-
-        let expectDiskCount = expectation(description: "Disk-counts change")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectDiskCount, inputChecker: { state in
-            return state.diskCount[.dark] == 13 && state.diskCount[.light] == 0
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'GameOver'")
-        let stateSubscriber3 = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .gameOver
-        })
-        gameState.subscribe(stateSubscriber3)
-        stateSubscriber3.store(in: &cancellables)
-        
-        if let onEnter = NextTurnPhase.onEnter(previousPhase: AnyPhase(PlacingDiskPhase(cellChanges: [][...]))) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectTurnChange, expectDiskCount, expectPhaseToBeWaitForPlayer], timeout: 3.0)
-    }
-    
-    func testPassPhase1() throws {
-        let board = try Board(restoringFromText: """
-            --------
-            --------
-            xooooooo
-            xxoooooo
-            xoxxoooo
-            -ooxoooo
-            --oooo--
-            --oooo--
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.manual, .manual])
-        state.diskCount[.dark] = 7
-        state.diskCount[.light] = 32
-        state.phase = AnyPhase(PassPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 「パス」表示の依頼が出る
-        let expectPassNotificationRequest = expectation(description: "Request for pass-notification")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectPassNotificationRequest, inputChecker: { state in
-            return state.passNotificationRequest != nil
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-        
-        if let onEnter = PassPhase.onEnter(previousPhase: AnyPhase(NextTurnPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPassNotificationRequest], timeout: 3.0)
-
-        // その依頼が完了するまでは、勝手に次みは進まずに
-        // そのままの状態でとどまる
-        let expectRequestNotChange = expectation(description: "Request for pass-notification continues")
-        expectRequestNotChange.isInverted = true
-        stateSubscriber.reset(expectRequestNotChange, inputChecker: { state in
-            return state.passNotificationRequest == nil || state.phase.kind != .pass
-        })
-        wait(for: [expectRequestNotChange], timeout: 1.0)
-    }
-    
-    func testPassPhase2() throws {
-        let board = try Board(restoringFromText: """
-            --------
-            --------
-            xooooooo
-            xxoooooo
-            xoxxoooo
-            -ooxoooo
-            --oooo--
-            --oooo--
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.manual, .manual])
-        state.diskCount[.dark] = 7
-        state.diskCount[.light] = 32
-        state.phase = AnyPhase(PassPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 「パス」表示の依頼が出る
-        var requestId = UniqueIdentifier.invalid
-        let expectPassNotificationRequest = expectation(description: "Request for pass-notification")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectPassNotificationRequest, inputChecker: { state in
-            if let request = state.passNotificationRequest {
-                requestId = request.requestId
-                return true
-            }
-            return false
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        
-        if let onEnter = PassPhase.onEnter(previousPhase: AnyPhase(NextTurnPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectPassNotificationRequest], timeout: 3.0)
-
-        // その依頼を完了すると、依頼は消えて、
-        // NextTurnフェーズに遷移する
-        let expectPassNotificationRequestDisappear = expectation(description: "Request disappears")
-        stateSubscriber1.reset(expectPassNotificationRequestDisappear, inputChecker: { state in
-            return state.passNotificationRequest == nil
-        })
-
-        let expectPhaseToBeNextTurn = expectation(description: "Phase transits to 'NextTurn'")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectPhaseToBeNextTurn, inputChecker: { state in
-            return state.phase.kind == .nextTurn
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        game.dispatch(.passDismissed(requestId: requestId))
-        wait(for: [expectPassNotificationRequestDisappear, expectPhaseToBeNextTurn], timeout: 3.0)
-    }
-    
-    func testGameOverPhase() throws {
-        let board = try Board(restoringFromText: """
-            --------
-            --------
-            ----x---
-            ---xxx--
-            --xxxxx-
-            ---xxx--
-            ----x---
-            --------
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.manual, .manual])
-        state.diskCount[.dark] = 13
-        state.diskCount[.light] = 0
-        state.phase = AnyPhase(GameOverPhase())
-
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // ターンがnilに変わる
-        let expectTurnChange = expectation(description: "Turn changed to nil")
-        let stateSubscriber = EventuallyFulfill<State, Never>(expectTurnChange, inputChecker: { state in
-            return state.turn == nil
-        })
-        gameState.subscribe(stateSubscriber)
-        stateSubscriber.store(in: &cancellables)
-
-        if let onEnter = GameOverPhase.onEnter(previousPhase: AnyPhase(NextTurnPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectTurnChange], timeout: 3.0)
-        
-        // そのまま待っていても、GameOverのまま
-        let expectGameOver = expectation(description: "Still game over")
-        expectGameOver.isInverted = true
-        stateSubscriber.reset(expectTurnChange, inputChecker: { state in
-            return state.phase.kind != .gameOver
-        })
-        wait(for: [expectGameOver], timeout: 1.0)
-    }
-    
-    func testReset1() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.manual, .computer])
-        state.diskCount[.dark] = 9
-        state.diskCount[.light] = 12
-        state.phase = AnyPhase(WaitForPlayerPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        if let onEnter = WaitForPlayerPhase.onEnter(previousPhase: AnyPhase(InitialPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        
-        // リセットが要求されると、リセット確認依頼が出る
-        var requestId = UniqueIdentifier.invalid
-        let expectResetConfirmation = expectation(description: "Request for reset-confirmation")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectResetConfirmation, inputChecker: { state in
-            if let request = state.resetConfirmationRequst {
-                requestId = request.requestId
-                return true
-            }
-            return false
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        
-        game.dispatch(.reset())
-        wait(for: [expectResetConfirmation], timeout: 3.0)
-
-        // リセットをキャンセルすると、リセット確認依頼が取り下げられる
-        // そのまま特に何も起こらない
-        let expectResetConfirmationDisappear = expectation(description: "Request disappears")
-        stateSubscriber1.reset(expectResetConfirmationDisappear, inputChecker: { state in
-            return state.resetConfirmationRequst == nil
-        })
-        
-        let expectKeepWaitForPlayer = expectation(description: "Keep 'WaitForPlayer' and current disk count")
-        expectKeepWaitForPlayer.isInverted = true
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectKeepWaitForPlayer, inputChecker: { state in
-            return state.phase.kind != .waitForPlayer || state.diskCount[.dark] != 9 || state.diskCount[.light] != 12
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        game.dispatch(.resetConfirmed(requestId: requestId, execute: false))
-        wait(for: [expectResetConfirmationDisappear], timeout: 3.0)
-
-        // そのまま特に何も起こらない
-        wait(for: [expectKeepWaitForPlayer], timeout: 1.0)
-    }
-
-    func testReset2() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.manual, .computer])
-        state.diskCount[.dark] = 9
-        state.diskCount[.light] = 12
-        state.phase = AnyPhase(WaitForPlayerPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        if let onEnter = WaitForPlayerPhase.onEnter(previousPhase: AnyPhase(InitialPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        
-        // リセットが要求されると、リセット確認依頼が出る
-        var requestId = UniqueIdentifier.invalid
-        let expectResetConfirmation = expectation(description: "Request for reset-confirmation")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectResetConfirmation, inputChecker: { state in
-            if let request = state.resetConfirmationRequst {
-                requestId = request.requestId
-                return true
-            }
-            return false
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        
-        game.dispatch(.reset())
-        wait(for: [expectResetConfirmation], timeout: 3.0)
-
-        // リセットを実行すると、リセット確認依頼が取り下げられてResetフェーズへ
-        let expectResetConfirmationDisappear = expectation(description: "Request disappears")
-        stateSubscriber1.reset(expectResetConfirmationDisappear, inputChecker: { state in
-            return state.resetConfirmationRequst == nil
-        })
-        
-        let expectPhaseToBeReset = expectation(description: "Phase transits to 'Reset'")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectPhaseToBeReset, inputChecker: { state in
-            return state.phase.kind == .reset
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        game.dispatch(.resetConfirmed(requestId: requestId, execute: true))
-        wait(for: [expectResetConfirmationDisappear, expectPhaseToBeReset], timeout: 3.0)
-    }
-    
-    func testResetPhase() throws {
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .light, playerModes: [.computer, .manual])
-        state.diskCount[.dark] = 9
-        state.diskCount[.light] = 12
-        state.phase = AnyPhase(ResetPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // いろいろリセットされる
-        let expectReset = expectation(description: "Reset!!!")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectReset, inputChecker: { state in
-            // リバーシ盤の更新依頼が出る
-            guard state.boardUpdateRequest != nil else { return false }
-            
-            // turnが黒になる
-            guard state.turn == .dark else { return false }
-            
-            // プレイヤーモードはどちらもmanualになる
-            guard state.playerModes[.dark] == .manual else { return false }
-            guard state.playerModes[.light] == .manual else { return false }
-            
-            // ディスク枚数は2, 2になる
-            guard state.diskCount[.dark] == 2 else { return false }
-            guard state.diskCount[.light] == 2 else { return false }
-
-            return true
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-
-        // また、次のフェーズは WaitForPlayer になる
-        let expectPhaseToBeWaitForPlayer = expectation(description: "Phase transits to 'WaitForPlayer'")
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectPhaseToBeWaitForPlayer, inputChecker: { state in
-            return state.phase.kind == .waitForPlayer
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        if let onEnter = ResetPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectReset, expectPhaseToBeWaitForPlayer], timeout: 3.0)
-    }
-    
-    func testResetOnThinkingPhase() throws {
-        ThinkingPhase.thinkingDuration = 2.0
-        
-        let board = try Board(restoringFromText: """
-            o-------
-            -o-o----
-            --oo-o--
-            --xoxox-
-            ---xox--
-            ---oxxx-
-            ---o-x--
-            --o-----
-            """)
-        var state = State(board: board, turn: .dark, playerModes: [.computer, .manual])
-        state.phase = AnyPhase(ThinkingPhase())
-        
-        let game = Game(state: state, middlewares: [Logger.self])
-        let gameState = game.statePublisher.share()
-
-        // 思考中になったらぐるぐるする
-        let expectThinking = expectation(description: "'thining' in state is true")
-        let stateSubscriber1 = EventuallyFulfill<State, Never>(expectThinking, inputChecker: { state in
-            return state.thinking
-        })
-        gameState.subscribe(stateSubscriber1)
-        stateSubscriber1.store(in: &cancellables)
-        
-        if let onEnter = ThinkingPhase.onEnter(previousPhase: AnyPhase(WaitForPlayerPhase())) {
-            game.dispatch(.thunk(onEnter))
-        }
-        wait(for: [expectThinking], timeout: 3.0)
-        
-        // リセット実行！
-        // 思考中ではなくなる。
-        // 勝手にディスクが打たれたりしない
-        gameState
-            .sink { state in
-                // リセット確認が来たら確認して実行する
-                if let request = state.resetConfirmationRequst {
-                    DispatchQueue.main.async {
-                        game.dispatch(.resetConfirmed(requestId: request.requestId, execute: true))
-                    }
-                }
-            }
-            .store(in: &cancellables)
-
-        let expectNotThinking = expectation(description: "'thinking' in state is false")
-        stateSubscriber1.reset(expectNotThinking, inputChecker: { state in
-            return !state.thinking
-        })
-        
-        let expectNotTransitToPlaceDiskPhase = expectation(description: "Phase should not transit to 'PlaceDisk'")
-        expectNotTransitToPlaceDiskPhase.isInverted = true
-        let stateSubscriber2 = EventuallyFulfill<State, Never>(expectThinking, inputChecker: { state in
-            return state.phase.kind == .placeDisk
-        })
-        gameState.subscribe(stateSubscriber2)
-        stateSubscriber2.store(in: &cancellables)
-
-        game.dispatch(.reset())
-        wait(for: [expectNotThinking], timeout: 3.0)
-
-        // 待っていても勝手にディスクは打たれない
-        wait(for: [expectNotTransitToPlaceDiskPhase], timeout: 3.0)
-    }
+    }    
 }
